+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/buttons/KeyboardButton.kt
@@ -1,13 +1,88 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.buttons
 
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
+import com.github.insanusmokrassar.TelegramBotAPI.types.*
+import kotlinx.serialization.*
+import kotlinx.serialization.internal.StringDescriptor
+import kotlinx.serialization.json.*
+
+@Serializable(KeyboardButtonSerializer::class)
+sealed class KeyboardButton {
+    abstract val text: String
+}
 
 @Serializable
-data class KeyboardButton(
-    val text: String,
-    @SerialName("request_contact")
-    val requestContact: Boolean? = null,
-    @SerialName("request_location")
-    val requestLocation: Boolean? = null
-)
+data class SimpleKeyboardButton(
+    override val text: String
+) : KeyboardButton()
+
+@Serializable
+data class UnknownKeyboardButton internal constructor(
+    override val text: String,
+    val raw: String
+) : KeyboardButton()
+
+@Serializable
+data class RequestContactKeyboardButton(
+    override val text: String
+) : KeyboardButton() {
+    @SerialName(requestContactField)
+    val requestContact: Boolean = true
+}
+
+@Serializable
+data class RequestLocationKeyboardButton(
+    override val text: String
+) : KeyboardButton() {
+    @SerialName(requestLocationField)
+    val requestLocation: Boolean = true
+}
+
+@Serializable
+data class RequestPollKeyboardButton(
+    override val text: String,
+    @SerialName(requestPollField)
+    val requestPoll: KeyboardButtonPollType
+) : KeyboardButton()
+
+@Serializer(KeyboardButton::class)
+internal object KeyboardButtonSerializer : KSerializer<KeyboardButton> {
+    override fun deserialize(decoder: Decoder): KeyboardButton {
+        val asJson = JsonElementSerializer.deserialize(decoder)
+
+        return when {
+            asJson is JsonPrimitive -> SimpleKeyboardButton(asJson.content)
+            asJson is JsonObject && asJson.getPrimitiveOrNull(requestContactField) != null -> RequestContactKeyboardButton(
+                asJson.getPrimitive(textField).content
+            )
+            asJson is JsonObject && asJson.getPrimitiveOrNull(requestLocationField) != null -> RequestLocationKeyboardButton(
+                asJson.getPrimitive(textField).content
+            )
+            asJson is JsonObject && asJson.getObjectOrNull(requestPollField) != null -> RequestPollKeyboardButton(
+                asJson.getPrimitive(textField).content,
+                Json.nonstrict.fromJson(
+                    KeyboardButtonPollType.serializer(),
+                    asJson.getObject(requestPollField)
+                )
+            )
+            else -> UnknownKeyboardButton(
+                when (asJson) {
+                    is JsonObject -> asJson.getPrimitive(textField).content
+                    is JsonArray -> ""
+                    is JsonPrimitive -> asJson.content
+                },
+                asJson.toString()
+            )
+        }
+    }
+
+    override fun serialize(encoder: Encoder, obj: KeyboardButton) {
+        when (obj) {
+            is RequestContactKeyboardButton -> RequestContactKeyboardButton.serializer().serialize(encoder, obj)
+            is RequestLocationKeyboardButton -> RequestLocationKeyboardButton.serializer().serialize(encoder, obj)
+            is RequestPollKeyboardButton -> RequestPollKeyboardButton.serializer().serialize(encoder, obj)
+            is SimpleKeyboardButton -> encoder.encodeString(obj.text)
+            is UnknownKeyboardButton -> JsonElementSerializer.serialize(encoder, Json.nonstrict.parseJson(obj.raw))
+        }
+    }
+}
+