+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/PollContent.kt
@@ -1,13 +1,15 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.message.content
 
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.Request
-import com.github.insanusmokrassar.TelegramBotAPI.requests.send.SendPoll
+import com.github.insanusmokrassar.TelegramBotAPI.requests.send.polls.SendPoll
+import com.github.insanusmokrassar.TelegramBotAPI.requests.send.polls.createRequest
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.ContentMessage
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MessageContent
 import com.github.insanusmokrassar.TelegramBotAPI.types.polls.Poll
+import com.github.insanusmokrassar.TelegramBotAPI.types.polls.RegularPoll
 
 data class PollContent(
     val poll: Poll
@@ -17,10 +19,8 @@ data class PollContent(
         disableNotification: Boolean,
         replyToMessageId: MessageIdentifier?,
         replyMarkup: KeyboardMarkup?
-    ): Request<ContentMessage<PollContent>> = SendPoll(
+    ): Request<ContentMessage<PollContent>> = poll.createRequest(
         chatId,
-        poll.question,
-        poll.options.map { it.text },
         disableNotification,
         replyToMessageId,
         replyMarkup